Demo adding an undock timer as described by RedSky
Add an undock timer described by RedSky at https://forum.botengine.org/t/adding-a-delay-to-undock/3805?u=viir

Thinking about it again, I found a simpler solution than explained in the first response: We can store only the last time we saw a neutral in the bot memory structure. With this memory, we can use the time as an entropy source to do the randomization.