Support EVE Online developers adding alarm sounds to their apps
Add notifications shim for local beep sounds to example apps.

The notifications functionality is currently work-in-progress.
For now, this wrapper helps app developers use the same API as is currently expected for the built-in notifications functionality.

One design goal is to support users customizing notifications quickly and independently of the app program code. To achieve this goal, we let users work with the data they are already familiar with: The status text can be seen during a live run and is also prominently displayed when inspecting a session in the devtools. Using the status text as the data source also ensures users can easily reuse their notification rules with any app.

Alarms are a special case at the moment. These alarms are usually implemented using the general notifications functionality in botengine. That notifications functionality is not documented at the moment because the implementation is not done yet.

So we take a different path today to implement an alarm sound.
To play the alarm sound, you derive a notification from the app status text. Mid-term, there is probably a graphical user interface to configure this. Today, you configure notifications by adding a function that takes the status text as input and returns that list of beeps.
As a reminder: The status text includes the parts that you returned from `decideNextStep` as well as `statusTextFromDecisionContext`. You can use the output of any of these to trigger a notification.

To include your notifications function, you switch your app configuration from the usual `EveOnline.AppFrameworkSeparatingMemory.processEvent` to `EveOnline.AppFrameworkSeparatingMemory.processEventWithNotificationsShim`
That variant takes one more field in the configuration record, named `notificationsBeepSequenceFromStatusText`