Adapt to scenario found by binshi this week
Adapt context menu functions to the scenario shared by binshi at https://forum.botengine.org/t/autopilot-bot-fails-to-jump-to-next-stargate/3839

The first element in the cascade was outdated but close enough to the UI node to start the cascade. In the linked scenario, starting a new cascade should help because the new instance will contain the required entry, based on the game's progress in the meantime.

Also, integrate this feedback into other example apps as well.