Illustrate user interface to explore the UI tree in memory readings
(Following the discussion at https://forum.botengine.org/t/abyssal-window/3848)

Have a closer look at the sample shared by artsevruk with session-2021-02-12T15-03-53-6366a4.zip
Use a precompiled version of the alternate UI for EVE Online to load one of the readings from the game client in that session and explore the UI tree.

In the linked thread, Artsevruk called this part of the UI "Abyssal window". Looking at the reading, CCP seems to call it `KeyActivationWindow`.