Plant seed for image-processing based bots 🌱
Evolve towards image-processing on the foundation of the app framework refined in the last months. (See the diagram 'data flow in bot architecture separating memory' for an overview)
Add image reading to a framework for EVE Online. Expand the autopilot example app to use that in the scenario shared by Victor last week at https://forum.botengine.org/t/cloaking-device-in-warp-to-0-bot/3917

In that thread, we got some samples of the game client and found that we want to check for a greenish glow on a module button.
The changes made here in the framework modules and the autopilot app show how to get the screenshot image data from the game client and then compute the greenness of the relevant part of the screenshot.

+ Merge the screenshot reading from the other example app into the volatile host implementation. Add reading the window client react to help apps navigate the overall screenshot.
+ Expand the framework modules to support apps querying for a list of rectangular-shaped subregions of the game client window.
+ Consider the coordinate system that apps work with so far: Have the framework module map the crop regions to extract from screenshots.
+ Also, remove the obsolete effect queue in the framework module.
+ Adapt the autopilot app to query for one pixel 20 pixels left of the center of each module button in the ship UI. Use the pixel value for each module button to compute how green it is. Display this info for all module buttons in the status text.
+ Finally, to support the scenario discovered by Victor last week, adapt the function responsible for maintaining ship module activation to integrate the new color-based indication.

On the names in the framework module code: The `1x1` part of names means it relates to the screenshot's representation with original resolution. It is common to apply binning for these applications, so a `2x2` scope will likely appear.