Adapt to scenario `2021-05-03T07-17-50-579350.zip`
For the motivating issue and discussion of the solution, see https://forum.botengine.org/t/issues-with-warp-to-0-bot/3987

Since 39d0ffea8c seems a suitable base to start from, I take a closer look at what 39d0ffea8c does in scenario `2021-05-03T07-17-50-579350.zip`.s
And inspecting event 36 reveals that the app is discarding the open context menu here:
https://github.com/Viir/bots/blob/47ea0f21b874fb3ae80baa3630485dcbf845c127/implement/applications/eve-online/eve-online-warp-to-0-autopilot/EveOnline/AppFrameworkSeparatingMemory.elm#L281-L282
To explain that branch in the context menu function in more detail: It assumes the observed context menu was not opened for the target UI element because it is too far away from that element.
The threshold for the acceptable distance to the distance is coded here:
https://github.com/Viir/bots/blob/47ea0f21b874fb3ae80baa3630485dcbf845c127/implement/applications/eve-online/eve-online-warp-to-0-autopilot/EveOnline/AppFrameworkSeparatingMemory.elm#L278

By increasing that threshold to 20, we can make the bot work in scenario `2021-05-03T07-17-50-579350.zip`.
After this change, the bot clicked on the menu entry.

To summarize: The problem was that that specific game client placed the context menu further away than in the sessions used for training before.

From a longer-term view, this case shows why we automate the adaption of the behavior based on app metrics. After a user has selected both metrics and scenarios according to their priorities, we automatically search the space of behaviors.