Support randomized limit for outgoing commands per village
Implement the randomization feature suggested by Drklord two weeks ago at https://forum.botlab.org/t/farm-manager-tribal-wars-2-farmbot/3038/294

+ Expand accepted values for the limit outgoing command limit setting to include intervals with a minimum and maximum value, analog to the setting for break duration.
+ Pick a random value from the configured interval at decision time.

Below is the idea as found in the linked forum post:

For the new parameter “limit-outgoing-commands-per-village” is it possible to allow a range to be accepted like the breaking system “ex. 3-4”. This way it doesn’t look strange if all village are exactly 40 outgoing. If there’s a range we can apply it’ll vary each village making it looks less likely a bot is being used.