Adapt formatting of descriptions in example to new user interfaces
Adapt the format in the descriptions of the examples to the markdown rendering deployed on the reactor today.
Since today's deployment to the reactor, the catalog entries at https://catalog.botlab.org and https://reactor.botlab.org/catalog show bots descriptions as markdown.

The new UI uses the implementation from https://github.com/dillonkearns/elm-markdown/pull/108 for the markdown parsing and rendering.