Implement EVE Online bot for 2021-10-05 scenario of chuckyone
Implement for the scenario shared by chuckyone at https://forum.botlab.org/t/how-to-add-audio-message-if-neutral-or-enemy-hits-local/93/24

From memory reading, we already have the region where the UI element for the pilot is. We can use that to decide where to check the colors.
I found this color in the screenshot you posted: red=5/255, green=34/255, blue=120/255.

With the current training data set, the classification itself is simple:

+ For each pixel, compare the color to the one we found in the sample to classify that pixel as blue or not blue.
+ If any of the pixels in the UI element for the pilot is blue, we consider the pilot a friend.

The colors can vary for several reasons. Therefore we do not check the color for equality but use a more inclusive check.
In the scenario you described above, it is enough to require the blue component to be more than twice as intense as green and red.