Fix failing browser downloads depending on working directory
Fix a problem that caused the download of the web browser component to fail depending on the working directory of the botlab client process: Use a download directory independent of the working directory.

Implement the approach from https://github.com/hardkoded/puppeteer-sharp/issues/376#issuecomment-402551142 to configure the download directory.

(The problem also happened when starting the botlab client from the Windows start menu at the working directory was `C:\Windows\System32\` in that case.)