Speed up the scanning of the map to find barbarian villages
Reduce the time spent searching barbarian villages: Expand the functionality to read village data for multiple coordinates in a single request.
Adapt the bot program to request map chunks of 10x10 coordinates instead of a single coordinate.
The previous version of the bot often spent multiple times more time on the first farming cycle.

A big revelation in the exploration for me: Using promises and the `async`/`await` Javascript features, we removed the dependency on the callback. We can probably refactor the rest of the bot to get rid of other dependencies on `____callback____` too.