Improve DevTools session inspection with client version 2022-04-24
Reduce memory usage for session inspection in DevTools:
A recent investigation with Mike revealed that a session from an EVE Online bot was impossible to load in DevTools, with about 30 thousand events and the archive weighing about 700 MB. The biggest contributor to memory usage in the previous version was keeping all decompressed files from the zip archive in memory.
These changes reduced the memory usage for larger sessions to less than ten percent:
Specialize implementation to read events from play sessions in DevTools to reduce memory usage at runtime: Do not retain all files in memory and introduce a cache for recently used files. Stop retaining intermediate representations of events in memory.
With this version, I loaded the session mentioned earlier successfully. The web UI for session inspection does not yet work well with this number of events as it always renders all of them in the timeline.

Compared to the version from March, this version also brings a dedicated button to export the bot program code in the session summary.