Fix task scheduling in common bot frameworks
Avoid starting superfluous tasks: Consolidate and simplify the implementation to track the task in progress to ensure we don't miss remembering any of those. Refactor the event processing and task tracking for readability and remove some redundant parts from the state.

On the relation between tasks from the bot and tasks from the framework:
For some events, it is easier to see that we should forward them to the bot immediately.
One example is bot-settings-changed: If the bot decides to finish the session in response, the response should reach the host for the same event, not later. The same applies to changes in the status text.
What about commands from the bot? If the bot says "take a screenshot" we need to wait for the completion of the framework setup (volatile process) first.
Therefore we queue tasks from the bot to forward them when we have completed the framework setup.

This change in the framework also introduces a new constraint: We do not any more forward tasks from the bot to the host as long as we are waiting for any task to complete.