Prepare introduction of binning to reduce runtime expenses
Prepare the interface to game windows for binning and refine the API structure of requests and responses.
Adapt an example app to help see if the binning yields correct aggregate pixel values: Expand the status text to support comparing pixel values in original and binned image representations.
Adapt structure of responses to support deriving metrics with passive programs using the same interface as bot programs: One request/response structure design goal is to support the derivation of (performance) metrics and notifications with additional passive programs: The agent events should contain all the info for these derivations.
   To avoid depending on knowledge about the commands that caused these events, we repeat the window ID and reading ID in the response.