Make the bot send effects to the game client to collect coins
Make the bot send effects for a drag gesture to the game client to collect a coin. Use a drag gesture because a click could trigger other actions and modal dialogs. The samples look like the interactive region is always well below the coin icon: Add an offset.
Before picking a coin to collect, filter out instances so close to the window border that the offset would make us click outside the game window. In other words, avoid those instances whose associated interactive region is too far outside of the game window.
Pick a random of the remaining coins to collect.