Support adding notification sounds to a bot via shim
In the future, notifications will be a feature in the botlab client and the reactor because users want a flexible and easy-to-use interface to adapt notifications. For example, a notification via a smartphone might have a different form than a notification on the same machine that runs the game client. Also, changing the bot program code should not be required to adapt notifications to a user's needs.

The shim added with this commit supports notifications even before they are available in the botlab client.
This commit also expands an example bot for EVE Online with a notification sound to demonstrate how to use the shim.