Begin the switch to the advanced graphical user interface 🚀🪂
Reduce the costs for future development of the graphical user interface in the BotLab client: Get rid of the complex, expensive UI framework and replace it with an Elm-based app using HTML, CSS, and Elm-UI for rendering.
The biggest improvement comes from eliminating the chaotic flow in WPF and switching to the virtual DOM with automatic diffing. Using HTML and CSS is accidental. We don't yet have an alternative.
To limit the size of atomic steps in this project, version 2022-09-08 only replaces the part of the GUI on the running bot branch (What you see after choosing "Run Bot"). The UIs to install and for DevTools is still in WPF land and will be rewritten later.
The plan for this commit was to limit changes only to translating the existing UX to the new platform and then implementing the user-facing improvements in later commits.
Especially in the later stages of the project leading to this version, I enjoyed refining some minor aspects so that UX is already slightly better than the previous one.
One of the things that were much easier to implement in the new framework was the navigation stack. One significant improvement for the user is that the form to enter a path to a bot source now maintains its state (on the stack) and still holds the same path when the user navigates back to it. Before, the user had to reenter that path every time after returning.
I also changed the wording on the buttons in the play session stage.