Fix the overview interaction issue discovered by Thomas Maya on Friday
Adapt to fix the issue discovered in 'session-recording-2022-12-09T12-32-56' (for example, in event 359): The mining bot tried to click on an overview entry and failed because other UI elements mostly occluded that entry. In that case, the overview was scrolled down by about one row high, so the column headers almost wholly occluded the topmost entry.
We can avoid such a failure by ignoring the almost fully occluded entry and selecting another.
This commit changes the frameworks for EVE Online to support implementing this kind of solution in bot programs: In the module to parse the user interface from the memory reading, we introduce a new field on the record to model UI nodes. This new field tracks the subregion of the display region that is visible (and, therefore, clickable). We also change the functions to interact with UI elements to use this new field when computing locations for mouse clicks.

In a simulation based on 'session-recording-2022-12-09T12-32-56', the new mining bot added with this commit successfully avoided the abovementioned issue. For example, in event 359, we see the mouse click location switch to the second overview entry.