Adapt to the new behavior of context menu in surroundings button
Adapt to the new context menu in the game client discovered by Dean yesterday.
In Dean's new game client, the menu on the surroundings button opened with two stages initially. The versions of the game client seen earlier opened at most one stage of the menu in response to a single input from the user.
This commit adapts the frameworks for the examples for EVE Online to account for the new behavior in the menu. We can no longer assume that the cascade's last open stage matches where we want to go. This commit adds some state in the framework to track which were probably opened by us.
See the original discussion at https://forum.botlab.org/t/mining-bot-not-working/4558