Add bot interface version 2023-01-17 to support screenshot-reading bots
Support screenshot reading bots with a dedicated interface: Extend the bot-host interface to support reading game client screenshots. Begin implementing pixel binning as requested by bots.
This commit adds functionality to the DevTools interface to inspect the reading events and individual crops ordered by the bot. The new views display the actual images of the crops and offer to export the crop images as PNG files.

This commit also adds a method for bots to request random bytes to the interface between bots and hosts.