Fix the reading of the 'OK' button in message boxes
Fix the issue discovered by Marcos at https://github.com/Viir/bots/issues/62
Make the bot see the 'OK' button again: Take a screenshot of the game client to find the buttons in each message box: Use the display region of the button group found in the memory reading of the message box to determine where to look for buttons. The game client sometimes shows more than one button on this group, so begin by scanning the whole width of the button group for the vertical edges of the button(s). Then scan the pixels in the center of each button to check which label the button has.

This commit switches the examples for EVE Online to version 2023-01-17 of the bot interface to integrate with the tooling for image-processing development. The DevTools now show the contents of the screenshots and support exporting to dedicated PNG files to make it easier to collect training data for bot development.