Adapt to the game client from session-recording-2023-01-26T00-37-54
Adapt to the behavior of the game client seen in session-recording-2023-01-26T00-37-54
Marcos Macedo did the setup of the game client and shared his observations at https://github.com/Viir/bots/pull/64#issuecomment-1404492655

The previous example program code was based on the assumption that the drone window would be visible at least as long as the ship UI.
In session-recording-2023-01-26T00-37-54, we see that assumption was wrong: The drone window disappeared two steps earlier than the ship UI.

This commit replaces the earlier assumption with the following assumption: When we dock, the ship UI disappears at most four readings/steps later than the drone window.
The number four is coded in the declaration `dockWhenDroneWindowInvisibleCount` and can be adjusted there if needed.

The effect of the change in the program code can be seen in event 1426 of session-recording-2023-01-26T00-37-54: In that event, the previous program code decided to stay docked, and the new program decided to undock.