Simplify reading from screenshots with bot interface version 2023-02-06
Simplify the interface to read screenshots from game client windows: Simplify this interface by removing the bots' explicit ordering of representation and crops of screenshots. By automating this process, we free bot developers from the need to specify explicitly the parts of the screenshot needed in the current step.

How it works for the user: When looking into a session recording in DevTools, we see that the recording does not contain all representations of the screenshot of the game client in all readings. This is most obvious when we look at the level of binning groups: The API offers three levels of detail of the screenshot of the game client, each corresponding to a level of binning of the pixels: one-to-one-pixel(original resolution), 2x2 binning, and 4x4 binning.
When looking into a session recording, we sometimes see an event with only a representation in 4x4 binning but none in 2x2 or original resolution. The reason is that the bot (and parallel running programs, e.g., for metrics/notifications) has looked only at pixels in the 4x4 binned representation. The BotLab client notices this and only stores that representation of pixels in that event. This way, we automatically retain everything we need for testing and building simulations while at the same time not wasting too many resources on unnecessarily detailed event data.

This reduction of event data is similar to the reduction sometimes explicitly performed for purposes like curating training data or anonymization.

The difference between version 2023-01-17 and version 2023-02-06 of the API lies in how we determine the subset of detail levels and crops required from a specific screenshot: In 2023-01-17, the bot developer had to specify the crops and detail levels explicitly. Starting with 2023-02-06, this is no longer necessary: The bot code simply queries the pixels as needed, and the BotLab client detects and tracks the dependency on specific pixels in specific detail levels. This shift of responsibility led to simplifying the bot program codes. The simplification is also visible in the changes in the examples for EVE Online in this commit.

Besides the simplification of image processing, this commit also simplifies the memory of the most recent readings in the examples for EVE Online: Instead of remembering the complete reading from the game client, we now extract the parts that could be of interest in future steps, including the ship module button tooltips and context menus. This reduction to relevant details also reduces the processing required when viewing or inspecting the bot state.