Improve readability in the framework code to parse UI from screenshot
In the frameworks for EVE Online, clean up the code to parse user interface details from the game client's screenshot.
Also, expand the functionality to parse buttons in message boxes to emit an error message when the pixel data is missing at the location of the button group in the message box.