Expand using context menu to consider occlusion of target element
Expand the framework for using the context menu to account for scenarios where the target element is partially occluded. Partial occlusion of the target means we might not click on the default projected location. In this commit, we change the part computing the distance of the context menu to consider the actual click location from the previous step instead of assuming the center of the target element.