Adapt reading of repair shop window to new game client
Adapt the reading of the repair shop window to the observations shared by Marcos at https://github.com/Viir/bots/issues/71

Below is an excerpt from the linked discussion:
-----
I started to look into the game client session-recording-2023-02-23T07-24-46. I agree. It seems like the individual buttons are not in the memory reading. So bots will have to look at the screenshot to read the buttons.
For the buttons in message boxes, they already do that. Since bots already read from screenshots, no change in interfaces is necessary. The main novelty here is the new labels on the buttons.