Parse cells texts of inventory items in the list view
Adapt to the structure of UI elements found in sample 'inventory_list_session-2023-02-25T20-05-35-event-53-eve-online-memory-reading.json'
Move code to associate texts found in cells of a list views with the corresponding header text into a dedicated separate function to be reused for overview window entries and inventory entries.
Implement the splitting of the text by tab tags (<t>) as found in the inventory list view.