Use a variant of the EVE Online volatile process with short response times
Transfer from https://github.com/Arcitectus/Sanderling/commit/cc21ccfdf949b82f4e3703f0e18519c64c522d10

Fix the problem of timeouts when searching the UI tree root address
Support scenarios where searching the root of the UI tree takes more than 60 seconds: Adapt to the timeout of 60 seconds in the webserver platform: Rewrite the interface to the volatile process to move the search for the UI tree root on a separate thread and respond right away to requests to search the UI tree root.
With the previous version, the webserver platform would emit a TimeoutException with the message "The app did not return a HTTP response within 60 seconds." if the search took longer for that client.