Make selecting bots easier in the GUI with client version 2023-04-13
Make it easier to select a bot by displaying recently used bots
Make it easier for users to load a bot by adding bots used in recent local play sessions to the list of offered bots in this view. Expand the GUI to load a list of local play sessions and then load the summaries for the most recent ones, including the bot.