Adapt to the scenario discovered by Levris today
Adapt to the scenario shared by Levris at https://github.com/Viir/bots/issues/79#issuecomment-1515193171

From the discussion:
When you say 'overview panel', I guess you mean that UI element in the screenshot you added.

That screenshot is cropped; a part is missing, but I guess this is the menu you get when clicking the surroundings button.
It is easy to adapt a bot to use the entries in there.
We can do this by starting with a popular mining bot that already uses the menu from the surroundings button. That way, we only have to adjust the path in the menu. Most people use a submenu that contains the string 'asteroid belts' when they expand that menu to arrive at the representation of their mining site.

As I understand your screenshot, your setup has a submenu called '*Anomalies*' in the same menu.
So our next experiment is replacing that string used to filter submenus and testing that on your game client.