Fix bug in recording and memory usage with client version 2023-05-29
Fix bug in storing agent events for bots using interface version 2023-02-06: Fix a bug that caused the storing of agent events to fail when the play sessions used a bot with interface version 2023-02-06.
Consequently, events from the storage queue could not be dequeued, which led to ever-increasing memory usage of the BotLab client process.
This version fixes the part of the compiler specific to bot interface version 2023-02-06 to correctly handle requests to compute the estimated serialized size of the app state.