Publish BotLab client version 2023-08-20
Fix processing tasks from the bots init function:
Fix a bug that caused the BotLab client to discard any tasks a bot requested in the 'init' function. This bug also affected the example bot for Elvenar. To fix this bug, this commit splits part of the 'ProcessAgentEvent' method into a separate 'ProcessAgentEventResponse' method and uses that with the response we get from the bot's init function. This way, the tasks from init now are properly forwarded.