Publish BotLab client version 2023-12-30
Fix a problem that caused long delays in the acquisition of input focus in some scenarios: This problem in earlier versions became apparent, for example, when running a single bot that did not request input focus for more than 30 seconds. In that case, the subsequent following request for input focus could be delayed for up to seven seconds. The concrete delay depended on timing as well as the presence of other unrelated processes happening to have open TCP ports in the searched range (two seconds timeout)