Publish version 2024-02-10 of the BotLab client
This release expands the graphical interface to mitigate the problem of setups failing because the WebView2 runtime was not installed.
Several reports (see forum threads) show that loading the graphical interface failed on some setups because the WebView2 runtime was not installed there. The WebView2 is usually automatically installed on every Windows 11 computer, but it can be missing on Windows 10.
This version introduces a check for this exact failure mode. It also displays a dedicated message and an installation guide if it turns out that the WebView2 runtime is not yet installed.