Optimize ship saving speed in EVE Online
If we bail because ship hitpoints are too low, offer a new shortcut using the overview windows. If the unload station or structure is visible in one of the overview windows, use that one to dock instead of docking to a random station.
For more background, see the discussion at <https://forum.botlab.org/t/the-mining-robot-cant-find-its-way-home/4922/4>

Excerpt from the linked discussion:
[...]
Overall, it might be better for `runAway` to first search all overview windows and check if the same station we use for unloading is visible. This approach will save time as it means we do not need to descend into the cascade of the surroundings menu.
[...]