Expand the agent API for simulating inputs on game clients
Previously, bots used a volatile process to simulate inputs on game clients. The agent interface version 2024-10-19 introduces a new task variant for simulating inputs. This commit updates popular example projects to use version 2024-10-19 of the agent interface.
To use the new bots, load them in the BotLab client version 2024-10-21 or newer. Older versions of the BotLab client will not run these new bots.
Bots using this new standard interface also make it easier to implement a generic function to detect if a bot is stuck and intervene to get it unstuck.