Fix bug in bot framework for EVE Online that caused larger session logs
Fix a bug in the bot framework for EVE Online that prevented the removal of unused screenshots from the play session recording when run in combination with the current BotLab client. The retention of the whole last reading from the game client in the state removed with this commit meant that we did not achieve convergence for reduced agent events in the session as quickly anymore. The event reduction reduces pixel data, so the retention of the whole reading in the state meant that the state would differ initially. Convergence was only achieved after completing the following reading from the game client. However, the BotLab client only searched for convergence up to three events. Therefore, it did not reach convergence anymore in practice.
This commit resolves the convergence complications by removing the general retention of superfluous parts of the reading from the game client. It only keeps the offset needed for the placement of pointer effects.