Publish BotLab client version 2025-04-23
Version 2025-04-23 improves the reliability of file exports initiated by users in the GUI:

Avoid problems observed with the file download integrated with WebView2/Chromium by replacing it with a custom file export via the container app. The new way of exporting files in the user interface aims to solve two problems: Users reported the download in the WebView2 failing with screenshots showing error messages like the following in the inline "Downloads" UI in the Chromium view:

> Couldn't download - Something went wrong. Try again, or contact your organization.

Another problem with the past export UI was that sometimes, users had trouble locating the exported file. That happens because the "Downloads" UI in Chromium does not ask the user for a destination location but places it in the "Downloads" directory configured for the current user account. The new export flow helps avoid this issue by explicitly asking the user to confirm a destination location. The current implementation uses the default popup window ("Save as" as seen in many apps) offered by the .NET API, which has different problems because that separate window can sometimes be out of view. A better (future) solution would be to integrate that "Save as" into the existing application window, hiding the other content until the user completes or cancels the file-saving dialog.