Wait longer before retry activating ship module
To account for the longer delays indication of activity on the of the ship module button as reported at <https://github.com/Viir/bots/issues/86>

A quote from that thread:

```
There seems to be a problem activating miners. after an asteroid is targeted, the bot attempts to start my miners. frequently the miner doesn't activate fast enough so the bot deactivates the miner and tries again. often this results in depletion of my capacitor, so it takes longer to activate the miner. I believe the delay after activating a miner and waiting for it the become active should be increased.
```