Support multiple sites for unloading ore in bot settings prompt
Expand the bot settings and functions to unload the mined ore to support not only one but multiple sites as the destination. Search the 'Locations' window to see if any entry matches the configured station or structure names. This way, we can reuse the same bot settings prompt in different solar systems.