Support multiple sites for hiding when running away
Expand the bot settings and functions to run away and hide from danger to support not only one but multiple sites as the destination. Search the 'Locations' window to see if any entry matches the configured station or structure names. This way, we can reuse the same bot settings prompt in different solar systems.